chore: consolidate auth code
