+++ gateway-framework/src/auth.rs
@@ -1,23 +1,19 @@
-pub mod api_keys;
-mod common;
-
 use std::{
     collections::{HashMap, HashSet},
     sync::Arc,
     time::Duration,
 };
 
-use anyhow::{bail, ensure};
+use anyhow::{anyhow, bail, ensure};
 use dashmap::DashMap;
 use ordered_float::NotNan;
+use serde::Deserialize;
+use serde_with::serde_as;
 use thegraph_core::types::{alloy_primitives::Address, SubgraphId};
 use tokio::{sync::watch, time::MissedTickBehavior};
 
-use self::api_keys::APIKey;
-
 #[derive(Clone, Debug, Default)]
 pub struct AuthSettings {
-    /// TODO: remove this field (should not be included in reporting)
     pub key: String,
     pub user: Address,
     pub authorized_subgraphs: Vec<SubgraphId>,
@@ -25,8 +21,10 @@ pub struct AuthSettings {
 }
 
 impl AuthSettings {
+    /// Check if the given subgraph is authorized. If the set of authorized subgraphs is empty, then
+    /// any subgraph is authorized.
     pub fn is_subgraph_authorized(&self, subgraph: &SubgraphId) -> bool {
-        common::is_subgraph_authorized(&self.authorized_subgraphs, subgraph)
+        self.authorized_subgraphs.is_empty() || self.authorized_subgraphs.contains(subgraph)
     }
 
     pub fn is_any_deployment_subgraph_authorized(&self, subgraphs: &[SubgraphId]) -> bool {
@@ -34,6 +32,30 @@ impl AuthSettings {
             .iter()
             .any(|subgraph| self.is_subgraph_authorized(subgraph))
     }
+}
+
+#[serde_as]
+#[derive(Clone, Debug, Default, Deserialize)]
+pub struct APIKey {
+    pub key: String,
+    pub user_address: Address,
+    pub query_status: QueryStatus,
+    #[serde_as(as = "Option<serde_with::TryFromInto<f64>>")]
+    #[serde(rename = "max_budget")]
+    pub max_budget_usd: Option<NotNan<f64>>,
+    #[serde(default)]
+    pub subgraphs: Vec<SubgraphId>,
+    #[serde(default)]
+    pub domains: Vec<String>,
+}
+
+#[derive(Clone, Copy, Debug, Default, Deserialize)]
+#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
+pub enum QueryStatus {
+    #[default]
+    Active,
+    ServiceShutoff,
+    MonthlyCapReached,
 }
 
 #[derive(Clone)]
@@ -52,6 +74,7 @@ impl AuthContext {
     /// should be handled.
     pub fn check(&self, token: &str, domain: &str) -> anyhow::Result<AuthSettings> {
         ensure!(!token.is_empty(), "missing API key");
+        parse_api_key(token).ok_or_else(|| anyhow!("malformed API key"))?;
 
         if let Some(rate_limiter) = &self.rate_limiter {
             if rate_limiter.above_limit(token) {
@@ -60,8 +83,140 @@ impl AuthContext {
             rate_limiter.increment(token);
         }
 
-        // For now, the only option is an API key.
-        api_keys::check(self, token, domain)
+        if self.special_api_keys.contains(token) {
+            return Ok(AuthSettings {
+                key: token.to_string(),
+                user: Address::default(),
+                authorized_subgraphs: vec![],
+                budget_usd: None,
+            });
+        }
+
+        // Note: holding watch::Ref for the rest of the function
+        let api_keys = self.api_keys.borrow();
+
+        let api_key = api_keys
+            .get(token)
+            .ok_or_else(|| anyhow::anyhow!("API key not found"))?;
+
+        match api_key.query_status {
+            QueryStatus::Active => (),
+            QueryStatus::ServiceShutoff => {
+                bail!("payment required for subsequent requests for this API key");
+            }
+            QueryStatus::MonthlyCapReached => {
+                bail!("spend limit exceeded for this API key");
+            }
+        }
+
+        ensure!(
+            is_domain_authorized(api_key.domains.as_slice(), domain),
+            "domain not authorized by user"
+        );
+
+        Ok(AuthSettings {
+            key: api_key.key.clone(),
+            user: api_key.user_address,
+            authorized_subgraphs: api_key.subgraphs.clone(),
+            budget_usd: api_key.max_budget_usd,
+        })
+    }
+}
+
+fn parse_api_key(token: &str) -> Option<[u8; 16]> {
+    if token.len() != 32 {
+        return None;
+    }
+    let mut buf = [0_u8; 16];
+    faster_hex::hex_decode(token.as_bytes(), &mut buf).ok()?;
+    Some(buf)
+}
+
+/// Check if the query origin domain is authorized.
+///
+/// If the authorized domain starts with a `*`, it is considered a wildcard domain. This means that
+/// any origin domain that ends with the wildcard domain is considered authorized.
+///
+/// If the authorized domains set is empty, all domains are considered authorized.
+pub fn is_domain_authorized<S: AsRef<str>>(authorized: &[S], origin: &str) -> bool {
+    fn match_domain(pattern: &str, origin: &str) -> bool {
+        if pattern.starts_with('*') {
+            origin.ends_with(pattern.trim_start_matches('*'))
+        } else {
+            origin == pattern
+        }
+    }
+
+    authorized.is_empty()
+        || authorized
+            .iter()
+            .any(|pattern| match_domain(pattern.as_ref(), origin))
+}
+
+#[cfg(test)]
+mod tests {
+    use alloy_primitives::hex;
+
+    use super::{is_domain_authorized, parse_api_key};
+
+    #[test]
+    fn parse_invalid_length_api_key() {
+        assert_eq!(parse_api_key("0123456789abcdef0123456789abcde"), None);
+    }
+
+    #[test]
+    fn parse_invalid_format_api_key() {
+        assert_eq!(parse_api_key("abcdefghijklmnopqrstuvwxyz123456"), None);
+    }
+
+    #[test]
+    fn parse_valid_api_key() {
+        assert_eq!(
+            parse_api_key("0123456789abcdef0123456789abcdef"),
+            Some(hex!("0123456789abcdef0123456789abcdef"))
+        );
+    }
+
+    #[test]
+    fn authorized_domains() {
+        let authorized_domains = [
+            "example.com",
+            "localhost",
+            "a.b.c",
+            "*.d.e",
+            "*-foo.vercel.app",
+        ];
+
+        let sub_cases = [
+            ("", false),
+            ("example.com", true),
+            ("subdomain.example.com", false),
+            ("localhost", true),
+            ("badhost", false),
+            ("a.b.c", true),
+            ("c", false),
+            ("b.c", false),
+            ("d.b.c", false),
+            ("a", false),
+            ("a.b", false),
+            ("e", false),
+            ("d.e", false),
+            ("z.d.e", true),
+            ("-foo.vercel.app", true),
+            ("foo.vercel.app", false),
+            ("bar-foo.vercel.app", true),
+            ("bar.foo.vercel.app", false),
+        ];
+
+        for (input, expected) in sub_cases {
+            assert_eq!(
+                expected,
+                is_domain_authorized(&authorized_domains, input),
+                "match '{input}'"
+            );
+            // check all authorized when authorized set is empty
+            assert!(is_domain_authorized(&[] as &[&str], input));
+        }
     }
 }
